support psutil in pex (#14)
psutil specifies a tag of cp36, and isn't considered compatible against a tag of cp39, making it unusable in pex (as well as anything that depends on it, such as ipykernel). However, as far as I can tell, it should be compatible -- to make it so, for CPython, indicate that we're compatible with all minor versions, similar to py* tags.